+++ packages/twenty-server/test/integration/graphql/suites/inputs-validation/filter-validation/constants/successful-filter-input-by-field-metadata-type.constant.ts
@@ -1,10 +1,10 @@
-import { isDefined } from 'class-validator';
-import { type FieldMetadataTypeToTest } from 'test/integration/graphql/suites/args-validation/types/field-metadata-type-to-test';
+import { type FieldMetadataTypeToTest } from 'test/integration/graphql/suites/inputs-validation/types/field-metadata-type-to-test';
 import {
   TEST_TARGET_OBJECT_RECORD_ID,
   TEST_UUID_FIELD_VALUE,
-} from 'test/integration/graphql/suites/args-validation/utils/setup-test-objects-with-all-field-types.util';
+} from 'test/integration/graphql/suites/inputs-validation/utils/setup-test-objects-with-all-field-types.util';
 import { FieldMetadataType } from 'twenty-shared/types';
+import { isDefined } from 'twenty-shared/utils';
 
 export const successfulFilterInputByFieldMetadataType: {
   [K in FieldMetadataTypeToTest]: {
