+++ packages/twenty-server/test/integration/graphql/suites/inputs-validation/filter-validation/constants/test-object-gql-fields.constant.ts
