+++ packages/twenty-server/test/integration/graphql/suites/inputs-validation/filter-validation/full-name-field-filter-input-validation.integration-spec.ts
@@ -1,15 +1,15 @@
-import { successfulFilterInputByFieldMetadataType } from 'test/integration/graphql/suites/args-validation/filter-validation/constants/successful-filter-input-by-field-metadata-type.constant';
-import { testGqlSuccessfulScenario } from 'test/integration/graphql/suites/args-validation/filter-validation/utils/test-gql-successful-scenario.util';
-import { testRestSuccessfulScenario } from 'test/integration/graphql/suites/args-validation/filter-validation/utils/test-rest-successful-scenario.util';
-import { destroyManyObjectsMetadata } from 'test/integration/graphql/suites/args-validation/utils/destroy-many-objects-metadata';
-import { setupTestObjectsWithAllFieldTypes } from 'test/integration/graphql/suites/args-validation/utils/setup-test-objects-with-all-field-types.util';
+import { successfulFilterInputByFieldMetadataType } from 'test/integration/graphql/suites/inputs-validation/filter-validation/constants/successful-filter-input-by-field-metadata-type.constant';
+import { testGqlSuccessfulScenario } from 'test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-gql-successful-scenario.util';
+import { testRestSuccessfulScenario } from 'test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-rest-successful-scenario.util';
+import { destroyManyObjectsMetadata } from 'test/integration/graphql/suites/inputs-validation/utils/destroy-many-objects-metadata';
+import { setupTestObjectsWithAllFieldTypes } from 'test/integration/graphql/suites/inputs-validation/utils/setup-test-objects-with-all-field-types.util';
 import { FieldMetadataType } from 'twenty-shared/types';
 
-const FIELD_METADATA_TYPE = FieldMetadataType.EMAILS;
+const FIELD_METADATA_TYPE = FieldMetadataType.FULL_NAME;
 const successfulTestCases =
   successfulFilterInputByFieldMetadataType[FIELD_METADATA_TYPE];
 
-describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
+describe(`Filter input validation - ${FIELD_METADATA_TYPE}`, () => {
   let objectMetadataId: string;
   let objectMetadataSingularName: string;
   let objectMetadataPluralName: string;
@@ -31,7 +31,7 @@ describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
     ]);
   });
 
-  describe('Gql filterInput - success', () => {
+  describe('Gql filter input - success', () => {
     it.each(
       successfulTestCases.map((testCase) => ({
         ...testCase,
@@ -50,7 +50,7 @@ describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
     );
   });
 
-  describe('Rest filterInput - success', () => {
+  describe('Rest filter input - success', () => {
     it.each(
       successfulTestCases.map((testCase) => ({
         ...testCase,
