+++ packages/twenty-server/test/integration/graphql/suites/inputs-validation/filter-validation/multi-select-field-filter-input-validation.integration-spec.ts
@@ -1,20 +1,20 @@
-import { failingFilterInputByFieldMetadataType } from 'test/integration/graphql/suites/args-validation/filter-validation/constants/failing-filter-input-by-field-metadata-type.constant';
-import { successfulFilterInputByFieldMetadataType } from 'test/integration/graphql/suites/args-validation/filter-validation/constants/successful-filter-input-by-field-metadata-type.constant';
-import { testGqlFailingScenario } from 'test/integration/graphql/suites/args-validation/filter-validation/utils/test-gql-failing-scenario.util';
-import { testGqlSuccessfulScenario } from 'test/integration/graphql/suites/args-validation/filter-validation/utils/test-gql-successful-scenario.util';
-import { testRestFailingScenario } from 'test/integration/graphql/suites/args-validation/filter-validation/utils/test-rest-failing-scenario.util';
-import { testRestSuccessfulScenario } from 'test/integration/graphql/suites/args-validation/filter-validation/utils/test-rest-successful-scenario.util';
-import { destroyManyObjectsMetadata } from 'test/integration/graphql/suites/args-validation/utils/destroy-many-objects-metadata';
-import { setupTestObjectsWithAllFieldTypes } from 'test/integration/graphql/suites/args-validation/utils/setup-test-objects-with-all-field-types.util';
+import { failingFilterInputByFieldMetadataType } from 'test/integration/graphql/suites/inputs-validation/filter-validation/constants/failing-filter-input-by-field-metadata-type.constant';
+import { successfulFilterInputByFieldMetadataType } from 'test/integration/graphql/suites/inputs-validation/filter-validation/constants/successful-filter-input-by-field-metadata-type.constant';
+import { testGqlFailingScenario } from 'test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-gql-failing-scenario.util';
+import { testGqlSuccessfulScenario } from 'test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-gql-successful-scenario.util';
+import { testRestFailingScenario } from 'test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-rest-failing-scenario.util';
+import { testRestSuccessfulScenario } from 'test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-rest-successful-scenario.util';
+import { destroyManyObjectsMetadata } from 'test/integration/graphql/suites/inputs-validation/utils/destroy-many-objects-metadata';
+import { setupTestObjectsWithAllFieldTypes } from 'test/integration/graphql/suites/inputs-validation/utils/setup-test-objects-with-all-field-types.util';
 import { FieldMetadataType } from 'twenty-shared/types';
 
-const FIELD_METADATA_TYPE = FieldMetadataType.RATING;
+const FIELD_METADATA_TYPE = FieldMetadataType.MULTI_SELECT;
 const failingTestCases =
   failingFilterInputByFieldMetadataType[FIELD_METADATA_TYPE];
 const successfulTestCases =
   successfulFilterInputByFieldMetadataType[FIELD_METADATA_TYPE];
 
-describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
+describe(`Filter input validation - ${FIELD_METADATA_TYPE}`, () => {
   let objectMetadataId: string;
   let objectMetadataSingularName: string;
   let objectMetadataPluralName: string;
@@ -36,7 +36,7 @@ describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
     ]);
   });
 
-  describe('Gql filterInput - failure', () => {
+  describe('Gql filter input - failure', () => {
     it.each(
       failingTestCases.map((testCase) => ({
         ...testCase,
@@ -56,7 +56,7 @@ describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
   });
 
   // TODO : Refacto-common - Uncomment this
-  describe('Rest filterInput - failure', () => {
+  describe('Rest filter input - failure', () => {
     it.each(
       failingTestCases.map((testCase) => ({
         ...testCase,
@@ -74,7 +74,7 @@ describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
     );
   });
 
-  describe('Gql filterInput - success', () => {
+  describe('Gql filter input - success', () => {
     it.each(
       successfulTestCases.map((testCase) => ({
         ...testCase,
@@ -93,7 +93,7 @@ describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
     );
   });
 
-  describe('Rest filterInput - success', () => {
+  describe('Rest filter input - success', () => {
     it.each(
       successfulTestCases.map((testCase) => ({
         ...testCase,
