+++ packages/twenty-server/test/integration/graphql/suites/inputs-validation/filter-validation/select-field-filter-input-validation.integration-spec.ts
@@ -1,17 +1,19 @@
-import { successfulFilterInputByFieldMetadataType } from 'test/integration/graphql/suites/args-validation/filter-validation/constants/successful-filter-input-by-field-metadata-type.constant';
-import { testGqlSuccessfulScenario } from 'test/integration/graphql/suites/args-validation/filter-validation/utils/test-gql-successful-scenario.util';
-import { testRestSuccessfulScenario } from 'test/integration/graphql/suites/args-validation/filter-validation/utils/test-rest-successful-scenario.util';
-import { destroyManyObjectsMetadata } from 'test/integration/graphql/suites/args-validation/utils/destroy-many-objects-metadata';
-import { setupTestObjectsWithAllFieldTypes } from 'test/integration/graphql/suites/args-validation/utils/setup-test-objects-with-all-field-types.util';
+import { failingFilterInputByFieldMetadataType } from 'test/integration/graphql/suites/inputs-validation/filter-validation/constants/failing-filter-input-by-field-metadata-type.constant';
+import { successfulFilterInputByFieldMetadataType } from 'test/integration/graphql/suites/inputs-validation/filter-validation/constants/successful-filter-input-by-field-metadata-type.constant';
+import { testGqlFailingScenario } from 'test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-gql-failing-scenario.util';
+import { testGqlSuccessfulScenario } from 'test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-gql-successful-scenario.util';
+import { testRestSuccessfulScenario } from 'test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-rest-successful-scenario.util';
+import { destroyManyObjectsMetadata } from 'test/integration/graphql/suites/inputs-validation/utils/destroy-many-objects-metadata';
+import { setupTestObjectsWithAllFieldTypes } from 'test/integration/graphql/suites/inputs-validation/utils/setup-test-objects-with-all-field-types.util';
 import { FieldMetadataType } from 'twenty-shared/types';
 
-const FIELD_METADATA_TYPE = FieldMetadataType.ARRAY;
-// const failingTestCases =
-//   failingFilterInputByFieldMetadataType[FIELD_METADATA_TYPE];
+const FIELD_METADATA_TYPE = FieldMetadataType.SELECT;
+const failingTestCases =
+  failingFilterInputByFieldMetadataType[FIELD_METADATA_TYPE];
 const successfulTestCases =
   successfulFilterInputByFieldMetadataType[FIELD_METADATA_TYPE];
 
-describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
+describe(`Filter input validation - ${FIELD_METADATA_TYPE}`, () => {
   let objectMetadataId: string;
   let objectMetadataSingularName: string;
   let objectMetadataPluralName: string;
@@ -33,27 +35,27 @@ describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
     ]);
   });
 
-  // describe('Gql filterInput - failure', () => {
-  //   it.each(
-  //     failingTestCases.map((testCase) => ({
-  //       ...testCase,
-  //       stringifiedFilter: JSON.stringify(testCase.gqlFilterInput),
-  //     })),
-  //   )(
-  //     `${FIELD_METADATA_TYPE} field type - should fail with filter : $stringifiedFilter`,
-  //     async ({ gqlFilterInput: filter, gqlErrorMessage: errorMessage }) => {
-  //       await testGqlFailingScenario(
-  //         objectMetadataSingularName,
-  //         objectMetadataPluralName,
-  //         filter,
-  //         errorMessage,
-  //       );
-  //     },
-  //   );
-  // });
+  describe('Gql filter input - failure', () => {
+    it.each(
+      failingTestCases.map((testCase) => ({
+        ...testCase,
+        stringifiedFilter: JSON.stringify(testCase.gqlFilterInput),
+      })),
+    )(
+      `${FIELD_METADATA_TYPE} field type - should fail with filter : $stringifiedFilter`,
+      async ({ gqlFilterInput: filter, gqlErrorMessage: errorMessage }) => {
+        await testGqlFailingScenario(
+          objectMetadataSingularName,
+          objectMetadataPluralName,
+          filter,
+          errorMessage,
+        );
+      },
+    );
+  });
 
-  // // TODO : Refacto-common - Uncomment this
-  // describe('Rest filterInput - failure', () => {
+  // TODO : Refacto-common - Uncomment this
+  // describe('Rest filter input - failure', () => {
   //   it.each(
   //     failingTestCases.map((testCase) => ({
   //       ...testCase,
@@ -71,7 +73,7 @@ describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
   //   );
   // });
 
-  describe('Gql filterInput - success', () => {
+  describe('Gql filter input - success', () => {
     it.each(
       successfulTestCases.map((testCase) => ({
         ...testCase,
@@ -90,7 +92,7 @@ describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
     );
   });
 
-  describe('Rest filterInput - success', () => {
+  describe('Rest filter input - success', () => {
     it.each(
       successfulTestCases.map((testCase) => ({
         ...testCase,
