+++ packages/twenty-server/test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-gql-failing-scenario.util.ts
