+++ packages/twenty-server/test/integration/graphql/suites/inputs-validation/filter-validation/uuid-field-filter-input-validation.integration-spec.ts
@@ -1,19 +1,20 @@
-import { failingFilterInputByFieldMetadataType } from 'test/integration/graphql/suites/args-validation/filter-validation/constants/failing-filter-input-by-field-metadata-type.constant';
-import { successfulFilterInputByFieldMetadataType } from 'test/integration/graphql/suites/args-validation/filter-validation/constants/successful-filter-input-by-field-metadata-type.constant';
-import { testGqlFailingScenario } from 'test/integration/graphql/suites/args-validation/filter-validation/utils/test-gql-failing-scenario.util';
-import { testGqlSuccessfulScenario } from 'test/integration/graphql/suites/args-validation/filter-validation/utils/test-gql-successful-scenario.util';
-import { testRestSuccessfulScenario } from 'test/integration/graphql/suites/args-validation/filter-validation/utils/test-rest-successful-scenario.util';
-import { destroyManyObjectsMetadata } from 'test/integration/graphql/suites/args-validation/utils/destroy-many-objects-metadata';
-import { setupTestObjectsWithAllFieldTypes } from 'test/integration/graphql/suites/args-validation/utils/setup-test-objects-with-all-field-types.util';
+import { failingFilterInputByFieldMetadataType } from 'test/integration/graphql/suites/inputs-validation/filter-validation/constants/failing-filter-input-by-field-metadata-type.constant';
+import { successfulFilterInputByFieldMetadataType } from 'test/integration/graphql/suites/inputs-validation/filter-validation/constants/successful-filter-input-by-field-metadata-type.constant';
+import { testGqlFailingScenario } from 'test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-gql-failing-scenario.util';
+import { testGqlSuccessfulScenario } from 'test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-gql-successful-scenario.util';
+import { testRestFailingScenario } from 'test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-rest-failing-scenario.util';
+import { testRestSuccessfulScenario } from 'test/integration/graphql/suites/inputs-validation/filter-validation/utils/test-rest-successful-scenario.util';
+import { destroyManyObjectsMetadata } from 'test/integration/graphql/suites/inputs-validation/utils/destroy-many-objects-metadata';
+import { setupTestObjectsWithAllFieldTypes } from 'test/integration/graphql/suites/inputs-validation/utils/setup-test-objects-with-all-field-types.util';
 import { FieldMetadataType } from 'twenty-shared/types';
 
-const FIELD_METADATA_TYPE = FieldMetadataType.SELECT;
+const FIELD_METADATA_TYPE = FieldMetadataType.UUID;
 const failingTestCases =
   failingFilterInputByFieldMetadataType[FIELD_METADATA_TYPE];
 const successfulTestCases =
   successfulFilterInputByFieldMetadataType[FIELD_METADATA_TYPE];
 
-describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
+describe(`Filter input validation - ${FIELD_METADATA_TYPE}`, () => {
   let objectMetadataId: string;
   let objectMetadataSingularName: string;
   let objectMetadataPluralName: string;
@@ -35,7 +36,7 @@ describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
     ]);
   });
 
-  describe('Gql filterInput - failure', () => {
+  describe('Gql filter input - failure', () => {
     it.each(
       failingTestCases.map((testCase) => ({
         ...testCase,
@@ -55,25 +56,25 @@ describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
   });
 
   // TODO : Refacto-common - Uncomment this
-  // describe('Rest filterInput - failure', () => {
-  //   it.each(
-  //     failingTestCases.map((testCase) => ({
-  //       ...testCase,
-  //       stringifiedFilter: JSON.stringify(testCase.restFilterInput),
-  //     })),
-  //   )(
-  //     `${FIELD_METADATA_TYPE} field type - should fail with filter : $stringifiedFilter`,
-  //     async ({ restFilterInput: filter, restErrorMessage: errorMessage }) => {
-  //       await testRestFailingScenario(
-  //         objectMetadataPluralName,
-  //         filter,
-  //         errorMessage,
-  //       );
-  //     },
-  //   );
-  // });
+  describe('Rest filter input - failure', () => {
+    it.each(
+      failingTestCases.map((testCase) => ({
+        ...testCase,
+        stringifiedFilter: JSON.stringify(testCase.restFilterInput),
+      })),
+    )(
+      `${FIELD_METADATA_TYPE} field type - should fail with filter : $stringifiedFilter`,
+      async ({ restFilterInput: filter, restErrorMessage: errorMessage }) => {
+        await testRestFailingScenario(
+          objectMetadataPluralName,
+          filter,
+          errorMessage,
+        );
+      },
+    );
+  });
 
-  describe('Gql filterInput - success', () => {
+  describe('Gql filter input - success', () => {
     it.each(
       successfulTestCases.map((testCase) => ({
         ...testCase,
@@ -92,7 +93,7 @@ describe(`Filter args validation - ${FIELD_METADATA_TYPE}`, () => {
     );
   });
 
-  describe('Rest filterInput - success', () => {
+  describe('Rest filter input - success', () => {
     it.each(
       successfulTestCases.map((testCase) => ({
         ...testCase,
