+++ packages/twenty-server/test/integration/graphql/suites/inputs-validation/utils/destroy-many-objects-metadata.ts
